Create Knowledge Management: Readable parent field value update through script
Requirement: To update the Knowledge article parent field. 
Challenges: If you have a readable field. It's difficult to update it. We are not able to edit the record by opening the form and list view.
Overcome: To overcome this problem, we can use this script to update the parent field.