Pull 24 3 - GlideRecordSecure Syntax Macro (#140)
* Create CreateRecord

GlideRecordSecure is a class inherited from GlideRecord that performs the same functions as GlideRecord, and also enforces ACLs.
This template will create a record only if user has permission

* Rename CreateRecord to CreateRecord.js

filename update

* Create updateRecordQuery.js

This script is designed to update multiple records in a specified table within ServiceNow using the GlideRecordSecure API to ensure update happen if user has permission to write

* Create CreateOrUpdate.js

This macro checks for the existence of a record in a specified table using a defined query. If the record is found, it updates a specified field with a new value. If the record is not found, it initializes a new record and inserts it into the table. Users should modify 'table_name', 'field_name', 'field_to_update', and the corresponding values as needed. The script logs the sys_id of the updated record or the ID of the newly created record for tracking purposes.

* Create deleteMultiple.js

this script provides an effective way to clean up records from a specified table based on defined conditions, ensuring efficient data management while respecting user permissions. It is particularly useful for maintaining the integrity and relevance of data in ServiceNow.

* Create updateMultipleTemplate.js

this script provides an efficient way to update multiple records in ServiceNow, allowing for streamlined data management while ensuring that user permissions are respected. Users should ensure that the encoded query is correctly defined to target the intended records.

* Create getRecord.js

this script provides a straightforward method for checking the existence of a record in ServiceNow and allows for further processing if the record is found. It is useful for scenarios where specific records need to be accessed and manipulated based on their unique identifiers.

* Create readme.md

Instruction and Implementation Details

* Delete GlideRecordSecure/readme.md

Removing your readme file, since this folder will not be specific to your commit, but open for others to add more files to, and as such, would be misleading. This is not a requirement in this repository, as it is not based on the same structure as code-snippets.

---------

Co-authored-by: Laszlo <47461634+Lacah@users.noreply.github.com>