Added two syntax macros for Adding attachment and sending outbound email notification through script. (#160)
* Create addAttachment.js

This syntax macro is used to create and add a new attachment to a specific record.

* Create sendOutboundEmail.js

The GlideEmailOutbound api can be used to trigger notification through script in case required to trigger notifications. The 'addRecipient' method can work in global scope(it is undocumented in servicenow) instead one can use setReplyTo for scopped app. Tested the snipped on my pdi before making submission.