Create GlideDateTime_DaysUntilFutureDate (#164)
* Create GlideDateTime_DaysUntilFutureDate

This syntax macro provides a reusable snippet for calculating the number of days between the current date and a specified future date. It initializes the current date, sets a future date, calculates the difference in milliseconds, converts the difference to days, and logs the result. This can be useful for deadline tracking, event planning, and other time-based calculations.

* Rename GlideDateTime_DaysUntilFutureDate to GlideDateTime/DaysUntilFutureDate

* Rename DaysUntilFutureDate to DaysUntilFutureDate.js

---------

Co-authored-by: Laszlo <47461634+Lacah@users.noreply.github.com>